queryrangebase definitions protos & aliases Signed-off-by: Owen Diehl <ow.diehl@gmail.com>
