+++ pkg/querier/queryrange/queryrangebase/definitions/definitions.proto
@@ -1,0 +1,24 @@
+syntax = "proto3";
+
+package definitions;
+
+import "github.com/gogo/protobuf/gogoproto/gogo.proto";
+
+option go_package = "github.com/grafana/loki/pkg/querier/queryrange/queryrangebase/definitions";
+option (gogoproto.marshaler_all) = true;
+option (gogoproto.unmarshaler_all) = true;
+
+// Defined here to prevent circular imports between logproto & queryrangebase
+message CachingOptions {
+  bool disabled = 1;
+}
+
+message PrometheusRequestHeader {
+  string Name = 1 [(gogoproto.jsontag) = "-"];
+  repeated string Values = 2 [(gogoproto.jsontag) = "-"];
+}
+
+message PrometheusResponseHeader {
+  string Name = 1 [(gogoproto.jsontag) = "-"];
+  repeated string Values = 2 [(gogoproto.jsontag) = "-"];
+}
