+++ pkg/querier/queryrange/queryrangebase/queryrange.proto
@@ -6,15 +6,12 @@ import "github.com/gogo/protobuf/gogoproto/gogo.proto";
 import "google/protobuf/any.proto";
 import "google/protobuf/duration.proto";
 import "pkg/logproto/logproto.proto";
+import "pkg/querier/queryrange/queryrangebase/definitions/definitions.proto";
 
 option go_package = "github.com/grafana/loki/pkg/querier/queryrange/queryrangebase";
 option (gogoproto.marshaler_all) = true;
 option (gogoproto.unmarshaler_all) = true;
 
-message PrometheusRequestHeader {
-  string Name = 1 [(gogoproto.jsontag) = "-"];
-  repeated string Values = 2 [(gogoproto.jsontag) = "-"];
-}
 message PrometheusRequest {
   string path = 1;
   int64 start = 2;
@@ -25,14 +22,10 @@ message PrometheusRequest {
     (gogoproto.nullable) = false
   ];
   string query = 6;
-  CachingOptions cachingOptions = 7 [(gogoproto.nullable) = false];
-  repeated PrometheusRequestHeader Headers = 8 [(gogoproto.jsontag) = "-"];
+  definitions.CachingOptions cachingOptions = 7 [(gogoproto.nullable) = false];
+  repeated definitions.PrometheusRequestHeader Headers = 8 [(gogoproto.jsontag) = "-"];
 }
 
-message PrometheusResponseHeader {
-  string Name = 1 [(gogoproto.jsontag) = "-"];
-  repeated string Values = 2 [(gogoproto.jsontag) = "-"];
-}
 
 message PrometheusResponse {
   string Status = 1 [(gogoproto.jsontag) = "status"];
@@ -42,7 +35,7 @@ message PrometheusResponse {
   ];
   string ErrorType = 3 [(gogoproto.jsontag) = "errorType,omitempty"];
   string Error = 4 [(gogoproto.jsontag) = "error,omitempty"];
-  repeated PrometheusResponseHeader Headers = 5 [(gogoproto.jsontag) = "-"];
+  repeated definitions.PrometheusResponseHeader Headers = 5 [(gogoproto.jsontag) = "-"];
 }
 
 message PrometheusData {
@@ -82,8 +75,4 @@ message Extent {
   reserved 3;
   string trace_id = 4 [(gogoproto.jsontag) = "-"];
   google.protobuf.Any response = 5 [(gogoproto.jsontag) = "response"];
-}
-
-message CachingOptions {
-  bool disabled = 1;
 }
